# src/main/resources/openapi/tft-service-api_v1.yaml
openapi: 3.0.1
info:
  title: TFT Service API
  description: ''
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: https://example.swagger.io/v2
tags:
  - name: summoner
    description: Operations about summoner
paths:
  /summoner/by-name/{summonerName}:
    get:
      tags:
        - summoner
      summary: Get summoner by summoner name
      operationId: getSummonerByName
      parameters:
        - name: summonerName
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Summoner'
        400:
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
  /summoner/by-id/{puuid}:
    get:
      tags:
        - summoner
      summary: Get summoner by summoner puuid
      operationId: getSummonerByPuuid
      parameters:
        - name: puuid
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Summoner'
        400:
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
  /summoner/by-name/{summonerName}/matches:
    get:
      tags:
        - summoner
      summary: Get list of match ids by summoner name
      operationId: getMatchesBySummonerName
      parameters:
        - name: summonerName
          in: path
          required: true
          schema:
            type: string
        - in: query
          name: count
          schema:
            type: integer
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummonerMatches'
        400:
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
  /summoner/by-id/{puuid}/matches:
    get:
      tags:
        - summoner
      summary: Get list of match ids by summoner puuid
      operationId: getMatchesBySummonerPuuid
      parameters:
        - name: puuid
          in: path
          required: true
          schema:
            type: string
        - in: query
          name: count
          schema:
            type: integer
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummonerMatches'
        400:
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
components:
  schemas:
    Summoner:
      type: object
      properties:
        summonerName:
          description: Summoner Name
          type: string
        summonerId:
          description: Encrypted summoner ID. Max length 63 characters.
          type: string
        puuid:
          description: Encrypted PUUID. Exact length of 78 characters.
          type: string
        summonerLevel:
          description: Summoner level associated with the summoner
          type: integer
          format: int64
        revisionDate:
          description: Date summoner was last modified specified as epoch milliseconds
          type: integer
          format: int64
    SummonerMatches:
      type: object
      properties:
        matchCount:
          description: Total match id retrieved from downstream
          type: integer
        matchIds:
          type: array
          items:
            type: string
    ServerError:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string